Finalize zsh config
This is going to be the last update to the zsh config
for the forseeable future.

This zsh config provides useful features such as autocompletion,
syntax highlighting and vi mode. The custom autocompletion and
the vi mode can for sure be very convenient, but they also form
habits that are of no in plain bash or zsh. The muscle memory that
we build up does not translate to other shells, where our
custom config is not in place.

Bash is de facto the standard shell on GNU/Linux distributions [1].
Furthermore it is the default shell for Fedora Kinoite, which is the
distribution that I have recently switched all of my systems over to.
Bash is usually available on any Linux based system, thus being
proficient in it provides much greater value long term than forming
muscle memory for a custom zsh environment.

All of that said I will be switching to (pretty much) plain bash and
we will see how that turns out. This zsh config should still be
functional in the future, as it does not depend on other plugin
managers. It only depends on three plugins, which all appear to be
very healthy open source projects, therefore I am confident that they
exist for a long time.

[1] https://unix.stackexchange.com/questions/28749/why-is-bash-standard-on-linux